pyproject.toml
[tool.poetry]
name = "djangodrf"
version = "0.1.0"
description = ""
authors = ["Aptemond <sofartem5@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.6"
python-decouple = "^3.8"
django-crispy-forms = "^2.0"
crispy-bootstrap4 = "^2022.1"
psycopg2 = "^2.9.9"
pillow = "^10.0.1"
djangorestframework = "^3.14.0"
django-filter = "^23.3"
djangorestframework-simplejwt = "^5.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
